requirements.txt
boto3>=1.26.0
google-auth>=2.16.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0
twilio>=7.16.0
aws-lambda-typing>=2.17.0
requests>=2.28.0